Supplies missing 'first_published_at' field to Document factory
The Edition factory attempts to set this through an evaluator association,
but this doesn't work if creating a Document directly rather than via an
Edition.
This change is needed prior to the next commit, where we'll be using the
field to determine which government the document should be linked with.