Add in history mode status banner
As part of the work to introduce history mode to content publisher we need to let publishers know when a document is in history mode. This introduces a notice that warns publishers when a political document is marked as being from a previous government. This feature is not yet ready for release and as such is hidden behind a feature flag. This will be removed at a later date.

Trello - https://trello.com/c/rhEtuy5V/1208-show-history-mode-status-on-summary-view

Co-authored-by: Bruce Bolt <bruce.bolt@digital.cabinet-office.gov.uk>