Consistent methods and tests for EditionHistory
To make state_event more intuitive bang and non-bang equivalents have
been created for all the methods that look up events.

As part of this overhaul I fixed a bug where next_event could cause a crash, as
index can return nil and nil + 1 = crash. I also amended the
edited_after_unpublishing? method to refer to an unpublishing so that it
would be easier to read, this has meant that some tests have needed
altering where history events are duplicates.

In the tests I set up a shared example for the bang methods so that we don't
have to write tests for each of them, it's not the easiest thing to read
but I think it's worth it for the brevity. I also expanded the data that
is in a number of tests to make it more explicitly test what the code is
wrote for.