Resync withdrawn editions with publishing-api
Also fixes the withdraw factory, so that it doesn't automatically create
an image revision and require to asset manage to be stubbed.

Also updates the `withdrawn_at` date in the factory to be an exact time
to remove differences in precision (which could be milliseconds) between
the date in the database and the one in the object being compared.