Associate generated document with the import record
Have used the :cascade strategy for when a document is deleted,
as we want to remove any associated Whitehall import records
if that happens.

It is not possible for a whitehall_import to have a document_id
of null.

We also want to create an index, for speed of lookup.