New table for imported assets
Uses :restrict foreign key relationship so that if whitehall_import
is deleted, an error is raised if it is associated with one or more
whitehall_imported_assets.
For images and attachments, uses :cascade, since if the associated
asset is deleted we'll want to delete the import record for it too.

Stores variants as an array, we can loop through these at the end
and delete them or set up redirects to the newly created asset.