WIP: figuring out how to do the processing of the asset
We have three options for dealing with image variants:
1. Delete
2. Redirect to the full size Content Publisher asset
3. Redirect to a similar sized Content Publisher asset

On the basis that someone may be hotlinking any asset, I'd say
redirecting is the best option. On the basis that it's an edge
case and we want to avoid unnecessary complications, I'd say
redirecting to the full size image is the easiest approach.