Create a module for BulkData
Bulk data is the namespace that we will be using for download large
amounts of data periodically for use in Content Publisher. Some examples
of candidates for this are topics, contacts and organisations. These are
data entities where we need to download all of the Publishing API's data
rather than a small sampling of on demand data.

Previously we have loaded this data on demand and it risks periodic
timeouts for the application. As this data can be loaded in numerous
places it also makes it difficult to determine performance issues in the
application as it may be a contributing factor to any other tasks we
think may be slow.

This module contains two common exceptions that we're anticipating to
use. LocalDataUnavailableError is for situations when we're expecting
data to be available in our cache but it is not present.
RemoteDataUnavailableError is for when we try to load remote data but
this fails.