BulkData cache singleton class
This class is intended to be the entry point for writing any BulkData.
It uses the Rails Redis cache as an underlying store.

Using Redis was a design choice for the following reasons:

- Redis offers caching across multiple GOV.UK boxes, so one write will
cover all machines
- Redis is fast to read and write from

And using the Rails cache implementation offered the following
advantages:

- Can be configured with [middleware][1] to avoid need to hit Redis multiple
  times in a request
- Redis namespacing
- Automatic serialisation/deserialisation

An important concern is that we need to ignore the guidance Rails
suggests to use a LRU redis cache, as we don't want to have keys
automatically cleared out of redis. Instead we just want them to expire
naturally.

The methods on this class allow adding an item to the cache, removing an
item, removing all items.

On top of this there are time based methods which are intended to be
used to determine when an item was added to cache. This is to enable
checking how old something is when deciding whether to add something or
not. This is a feature Redis doesn't have by default so it's quite
simple to just add this a different entry, the Rails cache entry
actually has some info about created time, but I have no idea how to
access it and also think it's probably nicer to not have to pull out the
whole cache entry to check it.

[1]: https://github.com/rails/rails/blob/9926fd86626690746c7ee78174c4d3577bdfa58d/activesupport/lib/active_support/cache/redis_cache_store.rb#L44