Add permissions to unwithdraw workflow
Adds manage_live_history_mode permissions to the unwithdraw workflow,
this follows the exact same rules as for withdrawing a document, so if a
document is in history mode and a user want to unwithdraw it they need
to have managing_editor and manage_live_history_mode permissions.

The managing_editor feature for unwithdrawing will be removed in future
work and replaced by request tests.