Adapt Government model for loading from Publishing API
This sets up this model to work with data from the Publishing API. It no
longer has class methods to look up the model as these are going to be
on a different class.

The input we expect from the Publishing API is a hash of base_path,
content_id, details, locale and title. We can look up started_on,
ended_on and current dates from the details hash.