GovernmentRepository class for BulkData
The purpose of this class is to mediate communication between Publishing
API and our cache adapter for governments.

It writes entries to the cache using a cache key that is defined as a
constant in this class. This is named as "government-v1" to indicate
that this is version 1 of the data we are caching. If we are to make any
backwards incompatible changes to the data we expect (such as adding an
extra mandatory field) then we will need to increment this number. This
will ensure that when the code change is deployed the old cache will not
be used anymore and will expire naturally.

It provides a number of methods to finding a particular government and
a method to populate the cache.