Set up spec helpers for bulk data
This clears the BulkData::Cache after every spec and helpers to populate
the cache before tests.

Default governments are set up for the most common case where we just
need to know about the current and previous governments.

It also provides helper methods for current and past government as these
are likely to be needed quite frequently.