Transition edition to use BulkData::GovernmentRepository
This creates instances of the GovernmentRepository class whenever we
need to load governments in the context of an edition. The factory has
been updated to reflect that governments can't be determined without an
external dependency and thus governments are always injected.

This does raise the prospect of some potential performance concerns for
the amount of classes initialised if we are frequently looking up
government information. Mostly though I expect these things to not occur
multiple times in the same request so I don't think it'll cause any
actual performance issues.