Helper method to run ActiveJobs exclusively
This cuts back on the boilerplate needed to run a worker job on a single
box and deny another boxes that try and run this at the same time. This
is a situation that occurs with the sidekiq scheduler where every box
with a sidekiq process tries to run the job at the same time and the
first one to start it wins.

We counter this with a PostgreSQL advisory lock which prevents any other
clients using the db while this lock is held (they gracefully exit).
Ideally we want to run this lock inside a transaction so that if
anything goes wrong the lock can be automatically closed, whereas if we
run it outside a transaction and something crashes we may need to
manually release the lock.