Add document live trait that sets first_published_at
The first_published_at value should always be set on documents that have
a live edition.

I didn't know a nice way to specify an optional trait to the association
method so have ended up with this array monstrosity. If you know a
better technique feel free to change it.