Minor refactoring of ResyncService specs
This moves the it block outside of a context to be before context as per
the more common convention. It also removes the government and political
tests from their draft and live context since they aren't resultant
effects of that particular context.

Finally it makes the draft and live tests more consistent with the other
edition state tests by creating an edition rather than a document by
factory bot.