Add job to populate bulk data asynchronously
This sets up a job that is run every 15 minutes to update the bulk data.
If accessing the data fails this job is retried.

The schedule is set with an interval of 15 minutes and a delay of 0s which
makes the job start immediately whenever the worker process is restarted. As
this job is run on a schedule it will execute on all machines at the
same time. To counter this the job is run exclusively using the method
on the parent class.