Asynchronously load bulk data if it is not available
This will schedule a background job to load the data if for some reason
it's not available in the current environment and a service has tried to
use it.

There is a risk that this could cause a spike in identical queued jobs
were a bunch of users to get this at the same time. This shouldn't
present much of a problem as the job only executes one at a time and
won't do anything if the data was updated in the last 5 minutes.