Add an error view for BulkData::LocalDataUnavailableError
If the application is ever in a state where there is not local data
unavailable it will return a 503 response and a simple error message. As
I anticipate that developers may often be people who see this message
(at least while sidekiq isn't run by default with the app) I thought it
was useful to have explicit instructions in a developer environment.

It was pretty hard to think of the best way to test this so I went with
this rather contrived example since documents_show is one of the main
routes. I expect that when the BulkData technique is used as part of
index page filters we can replace this test with a less contrived one.