Switch the parameters in filtering to match UI
Since the parameters from part of the UI in the form of the URL, it
makes sense that these follow the same terminology as is presented to
users.

This is a late change to bring this consistency in before the feature
launches and doing this later would be a backwards compatibility break.