Update side navigation component to require aria-label
Landmarks must have a unique role or role/label.
Given that we always have the main navigation on the page, any additional
navigation landmark must be accompanied by a label to make it unique.
The label is munged with the role which creates content that is being
picked up by screen readers when using landmarks navigation mode.