Configure Redis for resiliency and reporting
By default the RedisCacheStore uses a fault tolerant approach where it
returns nil if it can't access Redis and writes to a logger that we
haven't configured [1].

This isn't an ideal configuration for our use case as it's a critical
problem if the cache isn't available. To try resolve this we've got 3
reconnect attempts available if we don't manage to connect to Redis or
have lost our connection.

In the event that we have an error reading from the cache this will
report the error to sentry, which will hopefully help make problems
easier to diagnose.

[1]: https://github.com/rails/rails/blob/09a2979f75c51afb797dd60261a8930f84144af8/activesupport/lib/active_support/cache/redis_cache_store.rb#L472-L485