Don't report BulkData::RemoteDataUnavailableError on server errors
There's not a whole lot we can do when a dependent service is erroring
so lets not report it and wait to see if it causes problems in the app.

This is to resolve an issue where we start getting server errors
overnight in integration at the point of the data sync.