BulkData resilient to time existing but key not
I'm not sure how but we ended up in a situation on staging where the
created key was set but the cache entry was not. This causes the app to
fall into a loop where it will keep trying to run the job to populate
the cache and this job won't do anything because it thinks the cache is
populated.