Re-organise WhitehallImporter
sync and import are now called from import_and_sync function, this makes
the interface simpler from the rake tasks perspective and allows the
task to easily work out if the import has failed by asking
whitehall_import if the import has been completed. This wouldn't be
possible with the old implementation especially when we start adding
failures for the sync failing etc.

Makes use of the 'imported' status when the import part is complete and
moves chaging the status to completed to after the sync has run.

Using inspect also gives us more in depth error messages by returning
the error class as well as the message.