Quieten down Sidekiq for tests
It turns out that setting the active_job.logger to use Sidekiq was not
great for the test environment as all log entries were output to STDOUT.
This remedies this by setting the Sidekiq logger only in the production
and development environments, which allows the test environment to
continue to use the Rails default.

I initially tried to do this by changing the test environment file and
leaving dev and prod to use application.rb but this was harder to get
working since I'd have needed to initialise a new logger for this as the
Rails one had not yet been created.