Log the cause of BulkData::RemoteDataUnavailableError issues
As I learnt to my peril, this seems to happen more often than I'd have
liked. I saw times Publishing API was taking > 30 seconds to return this
simple list of 56 documents. I have no idea why it is taking that long
but it seems appropriate that we log something when we fail.