Set a huge timeout for loading governments
We've seen quite a few instances in the logs where requests to load
governments have taken absurdly long (up to 40 seconds). While this is
madness and should be resolved/understood (there are only 56
governments) we have the opportunity in Content Publisher to wait longer
for this data since it is loader asynchronously.

I've gone for a 30 second timeout as I expect Nginx will force a 504 at
around 15 seconds so this just has a good long buffer on that. This
replaces the default which seemed to timeout at 8 seconds.