Longer retry cycle for PopulateBulkDataJob
Previously this job would use the default amount of retries which is 5
with 3s between them. This had the effect that it was quite vulnerable
to reporting an error particularly in integration and staging during a
data sync.

Initially I thought we'd not want this to be retrying for more than 15
minutes as we run the job on a 15 minute interval. However I think it's
actually fine for a job to be still retrying for a lot longer after that
so then it only needs to tell us an error after a long time has passed.

Given it's now such a long time it seems reasonable to alert on any
error.