Pass 'record' (instead of 'whitehall_document') to Import class
We can access the whitehall document from the record object when
needed. Passing 'record' allows us to update the import record in
real time, e.g. updating the document ID associated with the record,
which we can only do once we've created the document. We will do
this in the next commit.

Have had to create a `:whitehall_import` factory and update the
tests.