Create new `whitehall_imported_assets` table
Stores variants as JSON, we can loop through these at the end
and delete them or set up redirects to the newly created asset.
JSON is in the form:
  { "s960": "https://.../.../s960_foo.jpg", "s420": "...etc" }

We'll later write a custom validator to ensure only one of
image_revision or file_attachment_revision is referenced by each
asset.