Allow ImageRevision and FileAttachmentRevision to receive record
Currently we don't act upon it, but there are a fair few code changes
involved in passing this extra parameter to these classes, so I've
encapsulated them in one commit for now.

In the next commit we'll change these classes to act upon the record.
And in the commit after that, we'll pass the record in create_revision
rather than passing 'nil'.