Validate that WhitehallImportedAsset only associated with 1 asset
We only want to be associated with one Image::Revision or one
FileAttachment::Revision. Prior to this commit, it was possible
to associate a single WhitehallImportedAsset with both.

Ideally this validation would happen at the database level, but
we can't create the Image::Revision/FileAttachment::Revision
ahead of time as we want to create the import record for it first,
in case something goes wrong. Therefore both fields are optional
at the database level.