Adds 'asset_manager_id' method to WhitehallImportedAsset
This follows the same convention as FileAttachment::Asset and
Image::Asset. In fact, the code has been copied from there and
tweaked slightly to reference the 'original_asset_url' key in the
database. We should probably consider extracting this out to a
shared place so that we don't have to maintain three methods.

This method will come in handy when we write the migration logic
in migrate_assets.rb.