Adds 'main_asset' method to WhitehallImportedAsset
Depending on if the imported asset is an image or a file attachment,
the way of determining the 'main' asset is different. (Image::Revision
requires an 'asset(960)' call, whereas FileAttachment::Revision
simply calls 'asset').
This decorator method removes that complexity, and will make it easier
and cleaner to continue to write migration logic in migrate_assets.rb.