Prevent importing the same assets multiple times
Images/attachments are allowed to have the same filenames - we
increment the name by 1 every time - but the preceding prefix
has to be different.

If an identical URL is encountered, we've already processed the
asset once and don't need to do it again.

Having this logic in place will make it easier to write the
redirect code for assets from Whitehall to Content Publisher,
where we have to consider whether an asset exists on live before
we can decide whether to delete it from the draft stack.