Adds 'content_publisher_asset' method to WhitehallMigration::AssetImport
This method determines which content publisher asset URL to return,
depending on whether it is a fileattachment or image revision, and
whether or not it is the canonical asset or a variant of it.

This decorator method will simplify the migration logic, as we can
simply check if 'content_publisher_asset' is nil to determine whether
or not the original whitehall asset should be redirected to it or
removed altogether.