Pass WhitehallMigration::DocumentImport to Import class
This parameter replaces the `whitehall_document` parameter, which
we can now access via the `document_import` parameter when needed.

Passing a reference to WhitehallMigration::DocumentImport allows
us to update the import record in real time, e.g. updating the
document ID associated with the record, which we can only do once
we've created the document. We will do this in the next commit.