Adds 'whitehall_asset_id' method to WhitehallMigration::AssetImport
This was copied from the 'asset_manager_id' method inside
FileAttachment::Asset and Image::Asset, and tweaked slightly to
reference the 'original_asset_url' key in the database.
However, I've named this 'whitehall_asset_id' as this is the ID
that Whitehall assigns the asset.

This method will come in handy when we write the migration logic.