Allow [Image/FileAttachment]Revision to receive `document_import`
`document_import` will be an instance of WhitehallMigration::DocumentImport.
Currently we pass `nil` in place of this, but as there are a fair
few code changes involved in the passing of this extra parameter to
these classes, I've encapsulated them in one commit for now.

In the next commit we'll change these classes to act upon the record.
And in the commit after that, we'll pass the record in create_revision
rather than passing 'nil'.