Compare editions with pub-api before resync
This code has been added to the Resync service because both live and
draft editions of the document need to be compared against they're
equivalent versions in publishing-api.

Collections publisher uses publishing-api to revert the contents of a
step-by-step to the published edition, by passing the "version" number
to the get_content call:
https://github.com/alphagov/collections-publisher/blob/22e0e234511610a66b8c6945a85d160da95c13a5/app/controllers/step_by_step_pages_controller.rb#L211

We can do something along those lines to make sure we're comparing the
correct version or edition.

As with Collections Publisher, the version number is gleaned from the
change history in publishing-api:
https://github.com/alphagov/collections-publisher/blob/22e0e234511610a66b8c6945a85d160da95c13a5/app/helpers/publishing_api_helper.rb#L4