Move components test into view tests
This view test wasn't in the rspec-rails directory of views and didn't
follow the conventions of rspec-rails view testing. This sets it up to
follow the expectations of
https://relishapp.com/rspec/rspec-rails/v/3-9/docs/view-specs/view-spec

The assertions are switched from the Rails test assert_select to the
Capybara have_selector methods to be more idiomatic for RSpec.