Use a tiny PNG fixture image by default
This creates a tiny (3kB) image that can be used by default when we
create an edition with an image. This replaces the jpg that was used
previously that was ~500kB. I had hoped this would speed up some of the
image tests but it didn't have much affect. However I think it's a net
benefit to still be working with a smaller image (logs smaller debugging
simpler) so felt it should go in.