Remove unnecessary route methods
This removes route naming for paths that already named in this file.
This is to bring more consistency to our routes file as this is the
precedent most routes follow. This approach is preferred to having the
same methods that produce identical results as it prevents the incorrect
method used in the wrong context.