Use patch routes for updating
This applies more consistency to our routes file so that any places
where we are updating data use PATCH routes rather than POST routes.
There is sometimes a bit of a grey area between these so I've tried to
make use of existing conventions already established in the app as much
as possible.