Write up testing strategy
We've struggled over time with a consistent approach to testing with
different understandings of our approaches. This document serves as an
attempt to put down our current thinkings. It's unlikely to be fully
exhaustive and I'm hoping it will be amended and improved over time.

I intend to follow this with a bunch of changes to our tests to increase
consistency.