Replace show document summary feature test with view test
This was a feature test that tested the rendering of a single view which
feels more suitable in a view test than a feature test. I'm not sure we
definitely need this test given it doesn't really test any logic but
didn't want to remove it without a replacement now since it is a key
view of the application - I did reduce the assertions as they didn't all
seem necessary.