Move format configuration tests to respective models
The format directory is something of an unusual one in Content
Publisher. It contains just configuration files which is rather strange
for a Rails app. The tests directory for this is a bit stranger since it
has a mix of feature tests and configuration tests.

This moves these configuration tests out of the spec/formats directory
and into the corresponding model directory.