Consolidate image feature testing
This does some work to reduce the feature tests around images (which are
the slowest tests we have) by removing dead end paths and only running the
tests with js when it is needed. There's still quite a few distinct
features we have regarding images so there is still a number of tests.