Add request tests for Documents
This adds requests test for the DocumentsController. It fixes a couple
of minor issues where we weren't asserting edition state and were using
an implicit redirect.

The feature tests that result in dead ends for users have been removed
as they are now covered by request test. As part of this we've lost an
assertion that the previous form input is shown to the user. This
is something that we could consider adding at request or view level (it
probably sits more naturally at view level) however it isn't something
tested on most other forms so felt like an anomaly and should be
considered as something to add as part of a broader strategy.