Add Capybara rspec matchers for request tests
This is added so that simple selector queries can be made as a result of
request tests. These aren't available by default for request tests but
are available in feature, view, controller, helper, and mailer tests.
The use of them is discouraged as we shouldn't be worrying too much
about HTML output of request tests but to me it doesn't make a huge
amount of sense to be missing a key assertion mechanism that we have in
other places. The use of this is also consistent with Rails'
assert_select that is used in many example of integration tests [1].

As far as I can tell, the reason RSpec Rails is not keen on these is
because request tests used to use the Capybara DSL which has been
deprecated, this is distinctly different from the RSpec Matchers since
they have a whole Capybara session.

[1]: https://guides.rubyonrails.org/testing.html#integration-testing