Require params rather than slicing and risking 500
In writing request tests for these interactors I became aware that if
either date or time was missing from the params then the application
would throw a 500. This can be resolved by requiring a params field
rather than slicing a hash, if the field is missing the app will then
raise an ActionController::ParameterMissing and the request will result
in a 400 bad request.

I felt that this didn't need explicit tests as it relies on default
Rails behaviour.