Submitted for review banner persists
This changes the submitted for review banner from being a temporary flash to
persisting while the document is in a `submitted_for_review` status so that
users can refer back to it and get the review link to share when they need to.

Since the banner is no longer a flash, we update the translation keys for the
text used by it to reflect this.