Change summary tab partial ordering
This reorders some of the partials on the summary tab. We have decided that
inset prompts, such as requirements or submitting for review, should appear at
the top as they are ephemeral and require action prior to publishing. Notices,
such as history mode and withdrawn, persist unless configuration is changed and
so should be next to content in the message hierarchy.