Add document history feature test
This adds a feature test for document history.

I think it makes sense to call document history a feature as you can now
interact with the page by navigating through each page of history thanks
to pagination. It also houses internal notes which in itself is a
feature. As well as the above, document history is only going to get
more complicated through potential improvements like filtering or search
if a user need is derived for these additions.

Internal notes feature has also been moved into the 'history' directory
as it feels more at home here than workflow due it not really being part
of the 'workflow' and living on the same page as document history.