Add defaults for 'topics' feature on formats
Previously we had to manually specify that topics were supported on
every format, which is unnecesary since they always are. This makes
having topics the default for a format, but inverts this for (most)
tests, in order to avoid having extra API stubs.

It seems reasonable to keep the 'topics' flag, since it represents a
'unit feature' of a format, even though it's technically mandatory.