Add fetch-mock node dependency
Jasmine lacks support for mocking fetch requests and only provides the
ability to mock XMLHttpRequests. To satisfy our needs to mock fetch
requests we can use the popular fetch-mock library.

This commit installs the dependency and adds a helper file to Jasmine to
load in the dependency and erase it's history after each example runs.