Re-order the modal-workflow-js methods
This moves the public methods to be defined before the ones that are
only used privately. I considered whether to prefix the methods that are
only used by this module with an underscore however we only seem to have
one example of that in this project so far so I wasn't sure if we'd
actually established that as a convention or not.