Add helper file for working with modal-dialogue in JS tests
In order to test modal functionality we need the ability to build and
remove the modal. This helper file builds the necessary HTML into the
DOM for the modal.

It'd be nicer if there was a way to pull this HTML directly from the app
rather than duplicating but I don't unfortunately know a way at the
moment.