Add tests for contact and video embed modals
This adds tests for these two classes that are used to run these modals.
These tests are very similar which reflects the code in the tested
classes is also very similar. After pondering for a while I came to the
conclusion that if you were to add more modals like these you'd probably
want to reduce the duplication in the underlying classes which may
itself present the opportunities to reduce the test duplication.