Update documents index pagination information
Wrong information was previously being displayed showing incorrect page
numbers for the text corresponding with the previous and next page links.

If we were on page two, the previous link would display page 2 and so
would the next link. Now the previous link points to page 1 and next (if
there is a next page) points to page 3.