Fix flakey array assertions
These tests were asserting that an array contained the expected items
and was in a certain order. This ordering is non-deterministic so we're
at risk of these causing failed builds as was seen in
https://ci.integration.publishing.service.gov.uk/job/content-publisher/job/modal-tests/4/console

The verify that these tests were actually flakey and that this fixed the
issue I ran these both 5000 times (I know that seems rather high, the
issue didn't seem to reveal itself on a 500 test run). The results were
as follows:

Before:

Finished in 5 minutes 51 seconds (files took 5.4 seconds to load)
10026 examples, 8294 failures

After:

Finished in 6 minutes 25 seconds (files took 5.73 seconds to load)
10026 examples, 0 failures