Move 'label' to i18n lib
The use of `t!` raises an exception if the translation does not
exist, so I had to amend the document_type factory to define this
I18n value on the fly (as the document_type `id` is randomly
generated for each test).

I did try removing the randomness and specifying a real document
type instead, but this caused an unprecedented number of issues.
Using a recognised document type means having to stub link/tag
associations in publishing API. It changes the intent_payload
`rendering_app` values. It breaks path prefixing. It breaks
document type comparison, as the label gets overwritten
by I18n rather than what is set in the factory. It tangentially
breaks a number of tests where we use an edition factory, as the
edition factory calls document_type_factory too.

Finally, there wasn't an obvious choice for specifying what kind
of document type to use for certain tests: many tests should work
irrespective of the document type and would lose readability if
they were now to be conflated with a specific document type ("does
this apply to all document types?", etc). On balance, dynamically
mocking the new I18n values we need feels like the best way forward.

The reason I've used `document_type.id` as the label is to make
the value unique, as otherwise some tests break due to 'ambiguous
matches found' when choosing a document type.

I've also had to create a `label` method, rather than replace calls
to `.label` in the views with `t` method calls, as there are so many
places that `.label` is used. We can look at making the i18n
retrieval consistent between fields in a later commit.