Move js-only class within the autocomplete component
The autocomplete component (being an enhanced input), as
 any other form element rely on `govuk-form-group` class to
generate consistent spacing across elements. This class
assumes the form elements are at a sibling level (and not
nested in unnecessary div wrappers) to style the last element
without spacing. This updates the way we apply the
`app-js-only` class to this instance of the component by applying the class to the component based on a `jsonly` flag.