Defines DocumentTypeHelper to allow fallback values for I18n keys
We need a mechanism that allows fallback values for I18n keys, such
that we can define a default I18n config for a document type and
child document types can override this where necessary.

I18n allows fallbacks, but in a not particularly DRY way. So, we
have defined our own helper class to deal with defaults. We did
consider naming it I18nHelper and defining a generic class that
implements the kind of defaulting we require, but we only see this
being used by DocumentType for the time being, so can be explicit
about its naming.