Remove check on multiple lead orgs
We now want to import editions which have multiple lead organisations
but use the 'primary' (lowest numerical lead_ordering) to assign to the
edition.

A check is already done here -
https://github.com/alphagov/whitehall/blob/master/app/controllers/admin/export/document_controller.rb#L45
to decipher if the lead organisation passed in matches the primary lead
org of the latest edition, this makes sure we are only importing
documents which the most current version is owned by the organisation
queried in whitehall.