Ignore users without a uid when importing users
When importing documents from Whitehall - and if the creator
of that document is not a Signon user - the import fails as
these Whitehall users do not have a uid.

We should not import users without a uid but we should create
events by these users.  Such documents should not have a user
assigned (i.e. no user is assigned).

Add very small performance change.