Fix flakey modal insertion
This fixes a test issue where the use of 'match' was mis-interpreting
the hyphen characters in a UUID as regex ranges, leading to a

  Regexp error:
     empty range in char class ...

Fixing this then exposed a timing issue, where the JS code had not yet
inserted the expected (contact) snippet into the body field. This adds
missing expectations to ensure we wait for the modal to close before
checking for the inserted snippet, as well as changing the modal
workflows to only close the modal when all its operations are finished.