Controller Reshuffle
Trello - https://trello.com/c/ZTRjP9eT/1099-break-up-the-actions-in-the-documentscontroller

The Documents controller has been largely unchanged since we introduced the concept of editions. This moves functionality from the documents controller into the editions controller and a newly created content controller to better fit with our data models. It also adjusts the copy and tests accordingly.