Add UID to test user seed data
There is an assumption that documents migrated from Whitehall - and those within
Content Publisher - are always associated with either a valid user (i.e. someone
who has a signon account and has actived that account by loggin in to a system)
or are un-assigned (nil), in the case of Whitehall 'system' users.  However,
this assumption is currently not implemented in the tests, where a 'test' user
is created without a UID.

This change re-enforces that assumption within the test suite for consistency.