Revert "Move 'label' to i18n lib"
It didn't feel right that we had an accessor method for just one field,
('label') where every other field has been workable within I18n. We
used the 'label' method for reading, and we used an I18n hack in the
tests to write, meaning we weren't reliably testing this behaviour.

We considered falling back to a default value for cases where the label
is not set, but this was a scenario we only envisaged happening in the
tests, so again wasn't the right approach.
The label is an inherent value of a document type and it's not too bad
for it to live closely with the rest of the document type data. We
weren't principally trying to internationalise the application, but to
use the I18n library to simplify our model. It doesn't make sense to
try to use I18n if it's causing more harm than the good it is doing.

This change frees us up to remove the I18n hack from our tests altogether.