Move tag i18n complexity out of model
We can use our new t_doctype method in the views directly instead.
I've had to change a few tests to be explicit about the tag IDs
they're using, as otherwise the label/hint can't be found even with
our defaulting mechanism. It felt cleaner to do this on a per-test
basis rather than setting a default tag ID in the factory.

This meant we no longer required the 'stringify_keys' behaviour
in the DocumentTypeHelper. We also no longer need to include the
helper inside the DocumentType model.