Refactor DocumentTypeHelper to be more concise
We now only use DocumentTypeHelper to retrieve explicit/default
values for fields, and not for anything higher up in the i18n
tree.
We can therefore simplify its API to take an edition (which can
be used to determine the document type ID) and a short substring
of i18n key which we will prefix the document type ID to.