Separate timeline entry into app partials
This has been put together to distinguish the needs of the Content
Publisher created timeline entries from those that are imported from
Whitehall. It was done the container class code was being applied to
only one of the 3 conditionals in this timeline_entry file.

This also resolves an issue where a view test was using a within which
has no effect (frustratingly - it'd be nice if it errored).

It's likely to be better if we have a component for a generic
timeline entry.