Use keyword args for all services
Previously we allowed a mixture of plain and keyword arguments for
classes in 'app/services'. This changes all services to use keyword
arguments only, noting the increase in verbosity in certain places
is balanced by the increased meaning, since we now say what each of
the arguments is (for). A good example of this is:

   job.arguments.first vs. edition_id: job.arguments.first