Test index filtering Publishing API down in view
The feature test that is being replaced is used to test a particular
conditional in the view. Our test strategy suggests that these should be
covered by view tests.

There is a change in the form to use an explicit route rather than
relying on Rails context for the route. Using context means that a view
cannot be tested in isolation, and in general I think it's a better
practice to name a route as it is otherwise unclear of it's destination
(plus the Rails magic routing is confusing).