Replace generic (body) fields with specific ones
This introduces a new "BodyField" class to represent the 'body' field
and its associated attributes. We want to start co-locating the various
behaviours (requirements, payload, etc.) of the 'body' field with its
data ("Tell-Don't-Ask"), and this class provides a starting point.

For the tests, I did explore building the field with a trait, like we
had previously. But FactoryBot doesn't seem to like the use of a trait
to switch the class. Instead of having a factory that just creates an
instance of a class, I thought it made more sense to do that manually.