Use BodyField to generate Publishing API payload
This moves the generation of the body-specific part of the payload
for an edition into the new BodyField class, which co-locates this
behaviour with the identity of the field. Rather than couple the
models to the stringified keys of the payload, we allow for shorter,
symbolic keys when we merge in the body-specific parto of the payload.

Since the contents of a document are not exclusive to the 'details'
part of the Publishing API payload (e.g. 'title'), and we anticipate
applying a similar change to other fields, it makes sense to merge
in the body-specific payload at the top-level.