Move requirements checks into BodyField class
This co-locates the requirements checks that are specific to the
body field as part of the class representing that field.

One surprise in doing this was that the GovspeakDocument interface,
which is used as part of the checks, expects an edition, which means
we have to pass one to the BodyField methods. In future work we could
explore adapting GovspeakDocument to accept a revision instead.