Rename govspeak(_input) to show it's body-specific
This renames the 'govspeak' and 'govspeak_input' fields to reflect
that they are specific to their use as the 'body' field. Since the
fields are used in two different contexts (contents/edit and
documents/show), it makes sense to move them nearer to their view.

In future we may want to explore view tests, but there is little
motivation to do this currently, since the 'body' field has extensive
coverage at the feature level.