Begin to make title/summary optional fields
This modifies the 'content/edit' and 'documents/show' views to pretend
that title/summary are optional fields, extracting the associated ERB
into field-specific partials, similarly to the 'body' field.

While there are other places where 'title' is shown to the user, these
two views are unique in purpose to modify and replay the value of the
'title' field; in other places, the 'title' is part of the layout.

In its use as a 'content' field, the 'title' is strongly coupled to the
'base_path', since the former determines the latter; hence the (re)naming,
which will become increasingly relevant in the following commits.