Fix 'translation missing' errors for supertype selection
Hint text is only defined for some supertypes, not others.
An attempt was made to only display hint text if it is not
'nil', however it is never 'nil' because the fallback value
is a string 'translation missing: en.supertypes.not-sure.hint'

This commit does an explicit check using 'exists?' and
removes a redundant 'hint' var that is only used in the
ternary.